backend/requirements.txt
Django>=5.0,<7.0
djangorestframework>=3.15
django-cors-headers>=4.0
dj-database-url>=2.0
whitenoise>=6.0
gunicorn>=22.0
psycopg2-binary>=2.9
Pillow>=10.0
openpyxl>=3.1
WeasyPrint>=62.0
